python_braille/requirements.txt
opencv-python
pytesseract
Pillow
pdf2image
fpdf
deep-translator
pyttsx3
pypiwin32
keyboard